Improve synchronization of projectiles and spells VFX
In an hypothetical scenario where we have hundreds of active projectiles, synchronizing all of their positions would be a bandwidth killer. For this reason and others, projectiles are now simulated on every client locally.
When a client shoots a projectile, that projectile is created immediately for that client. Then, a copy of that projectile, with the same starting position, velocity, direction and damage is created in every other client. Only the projectile simulated by the server will ultimately deal damage.